Foundation

Documentation

data App

The foundation datatype for your application. This can be a good place to keep settings and values requiring initialization before your application starts running, such as database connections. Every handler will have access to the data present here.

Constructors

App

Fields

appSettings :: AppSettings

appStatic :: Static

Settings for static file serving.

appConnPool :: ConnectionPool

Database connection pool.

appHttpManager :: Manager

appLogger :: Logger

data MenuItem

Constructors

MenuItem

Fields

menuItemLabel :: Text

menuItemRoute :: Route App

menuItemAccessCallback :: Bool

data MenuTypes

Constructors

NavbarLeft MenuItem

NavbarRight MenuItem

type Handler = HandlerT App IO

type Widget = WidgetT App IO ()

resourcesApp :: [ResourceTree String]

type Form x = Html -> MForm (HandlerT App IO) (FormResult x, Widget)

A convenient synonym for creating forms.

isAuthenticated :: Handler AuthResult

Access function to determine if a user is logged in.

unsafeHandler :: App -> Handler a -> IO a